Update tensor data types to float32 and ensure device compatibility in Nemotron and Parakeet models
Changed input audio features and tensor creation in NemotronH_Nano_VL_V2 to use float32 for improved compatibility. Updated tensor creation for video replacement token IDs and image context token IDs to ensure they are on the correct device. Additionally, disabled cuDNN for Conv2D layers in the Parakeet encoder to prevent initialization errors in subprocesses.

+++ vllm/model_executor/models/parakeet.py
@@ -54,13 +54,16 @@ class ProjectedParakeet(nn.Module):
             config, llm_hidden_size=llm_hidden_size, max_model_len=max_model_len
         )
         self.encoder = HFParakeetEncoder(self.config)
-        self.encoder = self.encoder.to(dtype)
+        self.encoder = self.encoder.to(torch.float32)
         self.projection = ParakeetProjection(self.config)
         self.projection = self.projection.to(dtype)
 
     def forward(
         self, input_features: torch.Tensor, attention_mask: torch.Tensor | None = None
     ) -> torch.Tensor:
+        # Disable cuDNN for the Conv2d subsampling layers in the Parakeet encoder.
+        # cuDNN fails to initialize in vLLM's EngineCore subprocess when running
+        # float32 convolutions; the PyTorch fallback path works correctly.
         with torch.backends.cudnn.flags(enabled=False):
             outputs = self.encoder(
                 input_features=input_features, attention_mask=attention_mask
